data/requirements.txt
pandas==2.0.3
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0